Refactor sidebar plugin
