--- src/Editor/Components/Placeholder.tsx
@@ -1,3 +1,0 @@
-export default function Placeholder() {
-  return <div className="placeholder">Enter some text...</div>;
-}
