--- src/Editor/Components/Print.tsx
@@ -1,16 +1,0 @@
-interface Props {
-  handleClick: () => void;
-}
-
-export default function Print({ handleClick }: Props) {
-  return (
-    <button
-      type="button"
-      aria-label="export file"
-      title="export file"
-      onClick={handleClick}
-    >
-      <span className="sidebar__icon icon__print" />
-    </button>
-  );
-}
